refactor(api): split plugin-instance projection out of audience handler
The audience HTTP handler had accreted the GET /admin/plugin-instances
endpoint — a ~220-line manifest→DTO projection that is not
audience-specific. It serves four surfaces (audience editor, trigger
picker #219, admin/plugins service badges, OAuth re-authorize banner
#228/#230), and its DTO was misleadingly named
`pluginInstanceForAudienceDTO`.

- Move ListPluginInstances + its DTOs into plugin_instance_handler.go.
  Receiver stays *AudienceHandler (it owns the store+snap deps and the
  route wiring), so router/main/tests are untouched — pure relocation.
- Rename pluginInstanceForAudienceDTO → pluginInstanceDTO, documenting
  all four consumers so future edits account for the shared surface.
- Fix a misleading CAS-conflict message on audience Update: it reused
  runstate's "run status transition lost to concurrent writer" string.
- Centralize the SQLite bool↔int64 convention in boolToInt64/int64ToBool
  helpers, replacing five scattered inline conversions.

audience_handler.go drops from 922 to ~660 lines, now focused on
audience CRUD. No behavior change; build + api/audience/dispatch/policy
test suites green.
